precision

+++ src/main/java/de/akquinet/jbosscc/guttenbase/tools/DropTablesTool.java
@@ -78,7 +78,10 @@ public class DropTablesTool
 
       for (final IndexMetaData index : table.getIndexes())
       {
-        statements.add("DROP INDEX " + schema + index.getIndexName() + ";");
+        if (!index.isPrimaryKeyIndex())
+        {
+          statements.add("DROP INDEX " + schema + index.getIndexName() + ";");
+        }
       }
     }
